backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
firebase-admin==6.5.0
numpy==1.26.4
scipy==1.13.0
reportlab==4.2.0
fpdf2==2.7.9
aiofiles==23.2.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0       # for tests
pytest==8.2.0
pytest-asyncio==0.23.6